+++ db.json
@@ -142,18 +142,14 @@
       "price": 220
     },
     {
-      "img": "src/img/887590c4.jpeg",
-      "title": "Аллегро",
-      "description": "салат Айсберг, бастурма, перець болгарський, помідори чері, фета, горіхи кеш'ю, соус",
-      "price": 110,
-      "id": "0AJRDp8"
-    },
-    {
-      "img": "src/img/887590c4.jpeg",
-      "title": "Аллегро",
-      "description": "салат Айсберг, бастурма, перець болгарський, помідори чері, фета, горіхи кеш'ю, соус",
-      "price": 110,
-      "id": "UecJT74"
+      "type": "pizza",
+      "id": "p3",
+      "img": "887590c4.jpeg",
+      "title": "anasda",
+      "description": "томат, моцарела, пармезан",
+      "size": "35",
+      "weight": "450",
+      "price": 220
     }
   ]
 }
