+++ src/components/UI/cards/DiscountCard.js
@@ -12,5 +12,3 @@ export const DiscountCard = (props) => {
         </div>
     );
 }
-
-// export default DiscountCard;
