+++ src/components/UI/carousel/SliderAlso.js
@@ -36,5 +36,3 @@ export const SliderAlso = () => {
         </Swiper>
     );
 };
-
-// export default SliderAlso;
