+++ src/components/UI/modal/ModalBasket.js
@@ -34,4 +34,3 @@ export const ModalBasket = () => {
     );
 }
 
-// export default ModalBasket;
